Updated .NET version and refactored service classes
Updated the `TargetFramework` in `BackgroundPlaygroundWorker.csproj` from `net6.0` to `net8.0` and upgraded several package references to newer versions. The base image in the `Dockerfile` has also been updated to `8.0`. Refactored `JokePollingService`, `JokesApiService`, and `ProtectedApiPollingService` classes to initialize certain fields directly in the constructor parameters. The `dotnetRunMessages` property in `launchSettings.json` has been changed from a string to a boolean. Lastly, the `version` in `global.json` has been updated to `8.0.204`.